Extract sign up form to component
