+++ app/routes/sign-up.js
@@ -5,6 +5,9 @@ import moment from 'moment';
 export default Route.extend(UnauthenticatedRouteMixin, {
   model() {
     return {
+      email: '',
+      password: '',
+      username: '',
       timeZone: moment.tz.guess()
     };
   },
